// src/pages/large-screen/wisdom-tourism/pages/tourist-analysis/types.ts
/**
 * @ Author: willysliang
 * @ CreateTime: 2024-10-19 13:20:58
 * @ Modifier: willysliang
 * @ ModifierTime: 2024-10-19 13:21:36
 * @ Description: TouristAnalysis 组件配置
 */

export interface ITouristDataProps {
  /** 旅游地 */
  address: string;
  /** 次数 */
  frequency: number;
  /** 人数 */
  people: number;
  /** 盈利 */
  profit: number;
}